refactor: 리듀서 덕 패턴 적용 및 폴더구조 변경
