--- src/lib/store/actionTypes.js
@@ -1,5 +1,0 @@
-export const ADD_BOARD = 'ADD_BOARD';
-
-export default {
-  ADD_BOARD,
-};
