+++ src/lib/store/reducer/boardReducer.js
@@ -1,5 +1,4 @@
-import { ADD_BOARD } from './actionTypes';
-import { getUniqueId } from '../utils/string';
+import { getUniqueId } from '../../utils/string';
 
 const makeBoard = (title = '') => ({
   id: getUniqueId(),
@@ -7,18 +6,20 @@ const makeBoard = (title = '') => ({
   items: [],
 });
 
+const ADD_BOARD_COLUMN_ACTION = 'BOARD/ADD_COLUMN';
+
 export const initialState = {
   boards: [makeBoard('할 일'), makeBoard('진행 중'), makeBoard('완료')],
 };
 
 export const addBoardAction = (title) => ({
-  type: ADD_BOARD,
+  type: ADD_BOARD_COLUMN_ACTION,
   payload: title,
 });
 
-const reducer = (state, action) => {
+const boardReducer = (state, action) => {
   switch (action.type) {
-    case ADD_BOARD:
+    case ADD_BOARD_COLUMN_ACTION:
       return {
         ...state,
         boards: [...state.boards, makeBoard(action.payload)],
@@ -28,4 +29,4 @@ const reducer = (state, action) => {
   }
 };
 
-export default reducer;
+export default boardReducer;
